gconf-2.m4 is needed for autoconf, bug #374561 by Piotrek Juzwiak. Also handle .la files removal better (bug #374573 by Piotrek Juzwiak). Remove old.
Package-Manager: portage-2.1.10.3/cvs/Linux x86_64

+++ media-libs/libcanberra/libcanberra-0.28-r2.ebuild
@@ -1,6 +1,6 @@
 # Copyright 1999-2011 Gentoo Foundation
 # Distributed under the terms of the GNU General Public License v2
-# $Header: /var/cvsroot/gentoo-x86/media-libs/libcanberra/libcanberra-0.28-r2.ebuild,v 1.2 2011/06/01 01:31:21 flameeyes Exp $
+# $Header: /var/cvsroot/gentoo-x86/media-libs/libcanberra/libcanberra-0.28-r2.ebuild,v 1.3 2011/07/10 18:44:50 pacho Exp $
 
 EAPI="4"
 
@@ -38,6 +38,11 @@ REQUIRED_USE="udev? ( alsa )"
 src_prepare() {
 	epatch "${FILESDIR}"/${P}-underlinking.patch
 
+	# gconf-2.m4 is needed for autoconf, bug #374561
+	if ! use gtk && ! use gtk3 ; then
+		cp "${FILESDIR}/gconf-2.m4" m4/ || die "Copying gconf-2.m4 failed!"
+	fi
+
 	eautoreconf
 	elibtoolize
 }
@@ -66,7 +71,9 @@ src_install() {
 
 	# Remove useless .la files
 	# la files in /usr/lib*/${P}/ are needed
-	rm -v "${ED}"/usr/$(get_libdir)/gtk-*/modules/*.la || die
+	if use gtk || use gtk3; then
+		rm -v "${ED}"/usr/$(get_libdir)/gtk-*/modules/*.la || die
+	fi
 	rm -v "${ED}"/usr/$(get_libdir)/*.la || die
 }
